refactor(upload): Move NewIssue to entities
